+++ CHANGELOG.md
@@ -15,8 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `factory`: Generates code for factories. Depends on `models`.
   - `queries`: Generates code for queries.
   - `dberrors`: Generates code for unique constraint errors. Depends on `models`.
-  - `joins`: Adds templates to the `models` package to generate code for joins e.g `models.SelectJoin.Table.LeftJoin.Rel`.
+  - `where`: Generates type-safe code for `WHERE` clauses in queries. Depends on `models`.
   - `loaders`: Adds templates to the `models` package to generate code for loaders e.g `models.SelectThenLoad.Table.Rel()`.
+  - `joins`: Adds templates to the `models` package to generate code for joins e.g `models.SelectJoin.Table.LeftJoin.Rel`.
 
 ### Changed